Dockerfile: Migrate distribution to Ubuntu
Ubuntu is a better choice than openSUSE for it's huge repositories and
user base which can decrease the usage of source compiled or manually
installed binaries.

This is just a basic migration which revises the setup to work with
Ubuntu and replace a few manually installed binaries with the ones
from the repositories.